Including Font-Awesome statically.
This is not ideal, but seems to be the best way to do it at the moment.
Using either `bower` or `component` (or some other package manager) doesn't
work, for two reasons:

1. `harp compile` fails on the LESS files that ship with FA. Not sure why.
2. It adds an extra step for the user after generating a slide deck.

We could link it directly from a CDN, but then your slides require an internet
connection, which is also less than ideal.

Font-Awesome is relatively stable, and it's not so big that I'm worried about
including it this way.